Allow the show vlan command without being privileged
Though the show running-config has to stay a privileged only command,
there had to be fall back on the default command processor.  Inheritance
seems the good solution.

The override of the delegate_to_sub_processor by the default command
processor was sort of a hack to make the enable processor exit directly
out of the switch without returning to its parent. An exception to
terminate the session was introduced instead of this hack to allow any
command processor to control the session termination.
