+++ fake_switches/arista/command_processor/default.py
@@ -11,8 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+from fake_switches.arista.command_processor import vlan_display_name
 from fake_switches.command_processing.base_command_processor import BaseCommandProcessor
+from fake_switches.command_processing.shell_session import TerminalExitSignal
 
 
 class DefaultCommandProcessor(BaseCommandProcessor):
@@ -23,11 +24,24 @@ class DefaultCommandProcessor(BaseCommandProcessor):
     def get_prompt(self):
         return self.switch_configuration.name + ">"
 
-    def delegate_to_sub_processor(self, line):
-        processed = self.sub_processor.process_command(line)
-        if self.sub_processor.is_done:
-            self.is_done = True
-        return processed
+    def do_exit(self):
+        raise TerminalExitSignal()
 
     def do_enable(self):
         self.move_to(self.enabled_processor)
+
+    def do_show(self, *args):
+        if "vlan".startswith(args[0]):
+            if len(args) == 2:
+                vlans = list(filter(lambda e: e.number == int(args[1]), self.switch_configuration.vlans))
+                if len(vlans) == 0:
+                    self.write_line("% VLAN {} not found in current VLAN database".format(args[1]))
+                    return
+            else:
+                vlans = self.switch_configuration.vlans
+
+            self.write_line("VLAN  Name                             Status    Ports")
+            self.write_line("----- -------------------------------- --------- -------------------------------")
+            for vlan in sorted(vlans, key=lambda v: v.number):
+                self.write_line("{: <5} {: <32} active".format(vlan.number, vlan_display_name(vlan)))
+            self.write_line("")
